refactor bank

+++ src/main/kotlin/br/com/finlegacy/api/features/userProfiles/domain/service/UserProfileServiceImpl.kt
@@ -1,11 +1,13 @@
 package br.com.finlegacy.api.features.userProfiles.domain.service
 
+import br.com.finlegacy.api.core.exceptions.BadRequestException
 import br.com.finlegacy.api.core.exceptions.ForbiddenException
 import br.com.finlegacy.api.core.exceptions.ItemNotFoundException
+import br.com.finlegacy.api.core.extensions.isValidId
 import br.com.finlegacy.api.core.result.Result
-import br.com.finlegacy.api.features.userProfiles.domain.model.UserProfileInfo
-import br.com.finlegacy.api.features.userProfiles.domain.model.UserProfileCreate
-import br.com.finlegacy.api.features.userProfiles.domain.model.UserProfileUpdate
+import br.com.finlegacy.api.features.userProfiles.domain.model.UserProfile
+import br.com.finlegacy.api.features.userProfiles.controller.dto.request.UserProfileCreateRequest
+import br.com.finlegacy.api.features.userProfiles.controller.dto.request.UserProfileUpdateRequest
 import br.com.finlegacy.api.features.userProfiles.domain.repository.UserProfileRepository
 import br.com.finlegacy.api.features.users.domain.model.User
 import br.com.finlegacy.api.features.users.domain.repository.UserRepository
@@ -14,8 +16,10 @@ class UserProfileServiceImpl(
     private val userRepository: UserRepository,
     private val userProfileRepository: UserProfileRepository
 ): UserProfileService {
-    override suspend fun findById(id: Long, uidLogged: String): Result<UserProfileInfo> {
+    override suspend fun findById(id: Long, uidLogged: String): Result<UserProfile> {
         return runCatching {
+            if (!id.isValidId()) throw BadRequestException("User Profile id")
+
             userProfileRepository.findById(id) ?: throw ItemNotFoundException("User Profile")
         }.fold(
             onSuccess = { Result.Success(it) },
@@ -23,7 +27,7 @@ class UserProfileServiceImpl(
         )
     }
 
-    override suspend fun findAll(uidLogged: String): Result<List<UserProfileInfo>> {
+    override suspend fun findAll(uidLogged: String): Result<List<UserProfile>> {
         return runCatching {
             userProfileRepository.findAll()
         }.fold(
@@ -37,6 +41,8 @@ class UserProfileServiceImpl(
             val userLogged:User = userRepository.findByUid(uidLogged) ?: throw ForbiddenException()
             if (!userLogged.userProfile.isSysAdmin) throw ForbiddenException()
 
+            if (!id.isValidId()) throw BadRequestException("User Profile id")
+
             if (userProfileRepository.delete(id)) {
                 true
             } else {
@@ -48,24 +54,29 @@ class UserProfileServiceImpl(
         )
     }
 
-    override suspend fun update(clinicUpdate: UserProfileUpdate, uidLogged: String): Result<UserProfileInfo> {
+    override suspend fun update(userProfileUpdate: UserProfileUpdateRequest, uidLogged: String): Result<UserProfile> {
         return runCatching {
             val userLogged:User = userRepository.findByUid(uidLogged) ?: throw ForbiddenException()
             if (!userLogged.userProfile.isSysAdmin) throw ForbiddenException()
 
-            userProfileRepository.update(clinicUpdate) ?: throw ItemNotFoundException("User Profile")
+            if (!userProfileUpdate.id.isValidId()) throw BadRequestException("User Profile id")
+            if (userProfileUpdate.name.isEmpty()) throw BadRequestException("User Profile name")
+
+            userProfileRepository.update(userProfileUpdate) ?: throw ItemNotFoundException("User Profile")
         }.fold(
             onSuccess = { Result.Success(it) },
             onFailure = { Result.Failure(it) }
         )
     }
 
-    override suspend fun create(clinicCreate: UserProfileCreate, uidLogged: String): Result<UserProfileInfo> {
+    override suspend fun create(userProfileCreate: UserProfileCreateRequest, uidLogged: String): Result<UserProfile> {
         return runCatching {
             val userLogged:User = userRepository.findByUid(uidLogged) ?: throw ForbiddenException()
             if (!userLogged.userProfile.isSysAdmin) throw ForbiddenException()
 
-            userProfileRepository.create(clinicCreate)
+            if (userProfileCreate.name.isEmpty()) throw BadRequestException("User Profile name")
+
+            userProfileRepository.create(userProfileCreate)
         }.fold(
             onSuccess = { Result.Success(it) },
             onFailure = { Result.Failure(it) }
